requirements.txt
# Core dependencies
flask>=2.0.0
flask-cors>=3.0.0
flask-limiter>=3.0.0
gunicorn>=20.1.0
python-dotenv>=0.19.0

# AI/ML
anthropic>=0.7.0
openai>=1.3.0

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.900

# Utilities
requests>=2.28.0
pyyaml>=6.0.0
python-dateutil>=2.8.2
argparse
colorama>=0.4.6
